Create Vega
Vega is a private university based in South Africa and is a member of the IIE. We are first year Bachelors in Computing and Game Development students and I think we can greatly benefit from this product. We use C# in Unity for game development. See website at https://www.vegaschool.com.